back_end/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.0.4
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
